move all management tools to the same folder
