+++ management-tools/components/CategoryTree.tsx
